command/jsonplan: updates to "config" section of output
It was determined that "source" for each output in "config" was going
to be unreasonably large, and as it isn't used *yet* we will leave it
out. This buys us some time to choose a more compact format - or to
change our minds and add it back in.

Also: fixed the "resources" block of "config.RootModule" to match the
RFC.